% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_fm.R
\name{fm_to_ng}
\alias{fm_to_ng}
\title{Femtomoles conversion for DNA}
\usage{
fm_to_ng(fm, dna_len)
}
\arguments{
\item{fm}{The amount of DNA in solution as fm}

\item{dna_len}{The length of DNA being used}
}
\value{
The amount of DNA in a solution as Nanograms
}
\description{
Convert Femtomoles units to Nanograms
}
\examples{
dna_ng <- fm_to_ng(8.8, 520);
dna_ng <- fm_to_ng(c(8.8, 10.1), c(520, 500));
dna_ng <- fm_to_ng(c(8.8, 10.1), 520);
}
