% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_fm.R
\name{fm_to_ng_table}
\alias{fm_to_ng_table}
\title{Femtomoles conversion table of DNA for experiments}
\usage{
fm_to_ng_table(fm, dna_len, required_fm)
}
\arguments{
\item{fm}{The amount of DNA in solution as fm}

\item{dna_len}{The length of DNA being used}

\item{required_fm}{The required amount of DNA in nanograms for the experimental method}
}
\value{
A table containing the fm, dna_len, DNA conc in ng, and DNA volume
}
\description{
Convert Femtomole units to Nanogram
}
\examples{
dna_fm <- ng_to_fm(8.8, 520, 250);
dna_fm <- ng_to_fm(c(8.8, 10.1), c(520, 500), 250);
dna_fm <- ng_to_fm(c(8.8, 10.1), 520, c(250,100));
}
