% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibson.R
\name{gibson}
\alias{gibson}
\title{Vector and Insert DNA volume calculations}
\usage{
gibson(v_dna_len, v_ng, i_dna_len, i_ng, ratio = 2, tot_dna_vol = 5)
}
\arguments{
\item{v_dna_len}{The length of the vector in base pairs (bp)}

\item{v_ng}{The mass (ng) of vector per ul}

\item{i_dna_len}{The length of insert in base pairs (bp)}

\item{i_ng}{The mass (ng) of insert per ul}

\item{ratio}{The ratio of insert to vector that you require. It is recommended to have a ratio of 2:1 or 3:1. Default is 2.}

\item{tot_dna_vol}{The total DNA volume used for the assembly. Default is 5 (for 10ul reactions)}
}
\value{
A table that includes the vector and insert final volumes and mass.
}
\description{
Calculates the amount of Vector and Insert DNA in required molar ratios. Uses the equation: required mass insert (g) = desired insert/vector molar ratio x mass of vector (g) x ratio of insert to vector lengths
}
\examples{
table <- gibson(4129, 200, 1787, 24.3)
table <- gibson(4129, 200, 1787, 24.3, 3)
table <- gibson(4129, 200, 1787, 24.3, 3, 10)
}
