% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_fm.R
\name{ng_to_fm}
\alias{ng_to_fm}
\title{Nanograms conversion for DNA}
\usage{
ng_to_fm(ng, dna_len)
}
\arguments{
\item{ng}{The amount of DNA in solution as ng, as outputted by nanodrop/qubit}

\item{dna_len}{The length of DNA being used}
}
\value{
The amount of DNA in a solution as Femtomoles
}
\description{
Convert Nanogram units to Femtomoles
}
\examples{
dna_fm <- ng_to_fm(8.8, 520);
dna_fm <- ng_to_fm(c(8.8, 10.1), c(520, 500));
dna_fm <- ng_to_fm(c(8.8, 10.1), 520);
}
