% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_fm.R
\name{ng_to_fm_table}
\alias{ng_to_fm_table}
\title{Nanograms conversion table of DNA for experiments}
\usage{
ng_to_fm_table(ng, dna_len, required_fm, total_vol)
}
\arguments{
\item{ng}{The amount of DNA in solution as ng, as outputted by nanodrop/qubit}

\item{dna_len}{The length of DNA being used}

\item{required_fm}{The required amount of DNA in femtomoles for the experimental method}
}
\value{
A table containing the ng, dna_len, DNA conc in fm, and DNA volume
}
\description{
Convert Nanogram units to Femtomoles
}
\examples{
dna_fm <- ng_to_fm(8.8, 520, 250);
dna_fm <- ng_to_fm(c(8.8, 10.1), c(520, 500), 250);
dna_fm <- ng_to_fm(c(8.8, 10.1), 520, c(250,100));
}
